Follow naming conventions of the audio API
Names of the packages and types should follow the names
used in https://platform.openai.com/docs/api-reference/audio
and be consistent with the naming strategy in other packages.

This is a breaking change but given the low current usage,
we are breaking it now before it's impossible to break.